C more functions largest prime factor commit

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-long long i, num;
+long i, num;
 num = 612852475143;
 
 for (i = 2; i < num; i++)
